+++ prpr/src/core.rs
@@ -12,10 +12,10 @@
 
 pub use macroquad::color::Color;
 
-pub const NOTE_WIDTH_RATIO_BASE: f32 = 0.13175016;
-pub const HEIGHT_RATIO: f32 = 0.83175;
+pub const NOTE_WIDTH_RATIO_BASE: f64 = 0.13175016;
+pub const HEIGHT_RATIO: f64 = 0.83175;
 
-pub const EPS: f32 = 1e-5;
+pub const EPS: f64 = 1e-5;
 
 pub type Point = nalgebra::Point2<f32>;
 pub type Vector = nalgebra::Vector2<f32>;
@@ -95,8 +95,8 @@ impl Default for Triple {
 }
 
 impl Triple {
-    pub fn beats(&self) -> f32 {
-        self.0 as f32 + self.1 as f32 / self.2 as f32
+    pub fn beats(&self) -> f64 {
+        self.0 as f64 + self.1 as f64 / self.2 as f64
     }
 }
 
@@ -104,7 +104,7 @@ impl Triple {
 pub struct BpmList {
     /// (beats, time, bpm)
     /// time in seconds
-    elements: Vec<(f32, f32, f32)>,
+    elements: Vec<(f64, f64, f64)>,
     /// cursor for searching, value is the index of `elements`
     cursor: usize,
 }
@@ -113,11 +113,11 @@ impl BpmList {
     /// Create a new BpmList from a list of (beats, bpm) pairs
     ///
     /// Basically just calculate the time for each pair(key frame)
-    pub fn new(ranges: Vec<(f32, f32)>) -> Self {
+    pub fn new(ranges: Vec<(f64, f64)>) -> Self {
         let mut elements = Vec::new();
         let mut time = 0.0;
         let mut last_beats = 0.0;
-        let mut last_bpm: Option<f32> = None;
+        let mut last_bpm: Option<f64> = None;
         for (now_beats, bpm) in ranges {
             if let Some(bpm) = last_bpm {
                 time += (now_beats - last_beats) * (60. / bpm);
@@ -130,7 +130,7 @@ impl BpmList {
     }
 
     /// Get the time in seconds for a given beats
-    pub fn time_beats(&mut self, beats: f32) -> f32 {
+    pub fn time_beats(&mut self, beats: f64) -> f64 {
         while let Some(kf) = self.elements.get(self.cursor + 1) {
             if kf.0 > beats {
                 break;
@@ -145,12 +145,12 @@ impl BpmList {
     }
 
     /// Get the time in seconds for a given `i + n / d`
-    pub fn time(&mut self, triple: &Triple) -> f32 {
+    pub fn time(&mut self, triple: &Triple) -> f64 {
         self.time_beats(triple.beats())
     }
 
     /// Get the beat coordinate for a given time in seconds
-    pub fn beat(&mut self, time: f32) -> f32 {
+    pub fn beat(&mut self, time: f64) -> f64 {
         while let Some(kf) = self.elements.get(self.cursor + 1) {
             if kf.1 > time {
                 break;
